chore: move `impl From<#name> for #cli_name` derive into submodule
